Rename visitor attributes to device attributes.
