+++ src/lib/dataElements/deviceAttributes.js
